Update README.md
# Mini-Project: ServeLink

ServeLink is an innovative service platform developed as a collaborative effort by a team of four. This MERN (MongoDB, Express.js, React.js, Node.js) stack project aims to bridge the gap between organizations and volunteers, creating a seamless connection for community service opportunities.

#  Key Features:
- Organization-Volunteer Matching:  Facilitates easy connection between service organizations and willing volunteers.
- MERN Stack Implementation:  Utilizes MongoDB, Express.js, React.js, and Node.js for a robust and scalable web application.
- MinIO Bucket Integration:  Incorporates MinIO for efficient object storage, enhancing data management capabilities.
- User-Friendly Interface:  Designed with both organizations and volunteers in mind for intuitive navigation and interaction.

# Current Status:
The project is fully functional but not yet publicly hosted. For a visual preview of the application's features and interface, please refer to the screenshots and descriptions in our README page.


This project represents our team's commitment to leveraging technology for social good, creating a platform that makes community service more accessible and efficient for all parties involved.